feat: add responses api
