+++ src/sparsevllm/entrypoints/openai/protocol/__init__.py
@@ -1,0 +1,2 @@
+"""OpenAI-compatible request and response protocol models."""
+
