+++ src/sparsevllm/entrypoints/openai/protocol/worker.py
@@ -1,0 +1,6 @@
+"""Worker control protocol placeholders.
+
+The worker endpoints currently expose JSON objects returned by the engine
+directly, so no request body models are needed here yet.
+"""
+
