+++ src/sparsevllm/entrypoints/openai/responses/__init__.py
@@ -1,0 +1,2 @@
+"""Responses API helpers."""
+
