+++ src/sparsevllm/entrypoints/openai/routes/__init__.py
@@ -1,0 +1,2 @@
+"""FastAPI routers for the worker OpenAI-compatible server."""
+
