+++ src/sparsevllm/entrypoints/openai/serving/__init__.py
@@ -1,0 +1,2 @@
+"""Endpoint serving helpers for the OpenAI-compatible worker server."""
+
